For all networking functions that currently take int32_t file descriptors, add versions that take intptr_t.  Remove manual marshalling of SafeHandle->int.
Once this package is published and consumed by all components, we can remove the int32_t versions of all of these functions.

+++ src/System.Net.Sockets/src/System/Net/Sockets/SocketAsyncContext.Unix.cs
@@ -263,23 +263,23 @@ namespace System.Net.Sockets
 
         private sealed class AcceptOperation : AcceptOrConnectOperation
         {
-            public int AcceptedFileDescriptor;
+            public IntPtr AcceptedFileDescriptor;
 
-            public Action<int, byte[], int, SocketError> Callback
+            public Action<IntPtr, byte[], int, SocketError> Callback
             {
-                private get { return (Action<int, byte[], int, SocketError>)CallbackOrEvent; }
+                private get { return (Action<IntPtr, byte[], int, SocketError>)CallbackOrEvent; }
                 set { CallbackOrEvent = value; }
             }
 
             protected override void Abort()
             {
-                AcceptedFileDescriptor = -1;
+                AcceptedFileDescriptor = (IntPtr)(-1);
             }
 
             protected override bool DoTryComplete(SocketAsyncContext context)
             {
                 bool completed = SocketPal.TryCompleteAccept(context._socket, SocketAddress, ref SocketAddressLen, out AcceptedFileDescriptor, out ErrorCode);
-                Debug.Assert(ErrorCode == SocketError.Success || AcceptedFileDescriptor == -1, $"Unexpected values: ErrorCode={ErrorCode}, AcceptedFileDescriptor={AcceptedFileDescriptor}");
+                Debug.Assert(ErrorCode == SocketError.Success || AcceptedFileDescriptor == (IntPtr)(-1), $"Unexpected values: ErrorCode={ErrorCode}, AcceptedFileDescriptor={AcceptedFileDescriptor}");
                 return completed;
             }
 
@@ -581,7 +581,7 @@ namespace System.Net.Sockets
             return true;
         }
 
-        public SocketError Accept(byte[] socketAddress, ref int socketAddressLen, int timeout, out int acceptedFd)
+        public SocketError Accept(byte[] socketAddress, ref int socketAddressLen, int timeout, out IntPtr acceptedFd)
         {
             Debug.Assert(socketAddress != null, "Expected non-null socketAddress");
             Debug.Assert(socketAddressLen > 0, $"Unexpected socketAddressLen: {socketAddressLen}");
@@ -590,7 +590,7 @@ namespace System.Net.Sockets
             SocketError errorCode;
             if (SocketPal.TryCompleteAccept(_socket, socketAddress, ref socketAddressLen, out acceptedFd, out errorCode))
             {
-                Debug.Assert(errorCode == SocketError.Success || acceptedFd == -1, $"Unexpected values: errorCode={errorCode}, acceptedFd={acceptedFd}");
+                Debug.Assert(errorCode == SocketError.Success || acceptedFd == (IntPtr)(-1), $"Unexpected values: errorCode={errorCode}, acceptedFd={acceptedFd}");
                 return errorCode;
             }
 
@@ -615,7 +615,7 @@ namespace System.Net.Sockets
 
                     if (isStopped)
                     {
-                        acceptedFd = -1;
+                        acceptedFd = (IntPtr)(-1);
                         return SocketError.Interrupted;
                     }
 
@@ -629,7 +629,7 @@ namespace System.Net.Sockets
 
                 if (!operation.Wait(timeout))
                 {
-                    acceptedFd = -1;
+                    acceptedFd = (IntPtr)(-1);
                     return SocketError.TimedOut;
                 }
 
@@ -639,7 +639,7 @@ namespace System.Net.Sockets
             }
         }
 
-        public SocketError AcceptAsync(byte[] socketAddress, int socketAddressLen, Action<int, byte[], int, SocketError> callback)
+        public SocketError AcceptAsync(byte[] socketAddress, int socketAddressLen, Action<IntPtr, byte[], int, SocketError> callback)
         {
             Debug.Assert(socketAddress != null, "Expected non-null socketAddress");
             Debug.Assert(socketAddressLen > 0, $"Unexpected socketAddressLen: {socketAddressLen}");
@@ -647,17 +647,17 @@ namespace System.Net.Sockets
 
             SetNonBlocking();
 
-            int acceptedFd;
+            IntPtr acceptedFd;
             SocketError errorCode;
             if (SocketPal.TryCompleteAccept(_socket, socketAddress, ref socketAddressLen, out acceptedFd, out errorCode))
             {
-                Debug.Assert(errorCode == SocketError.Success || acceptedFd == -1, $"Unexpected values: errorCode={errorCode}, acceptedFd={acceptedFd}");
+                Debug.Assert(errorCode == SocketError.Success || acceptedFd == (IntPtr)(-1), $"Unexpected values: errorCode={errorCode}, acceptedFd={acceptedFd}");
 
                 if (errorCode == SocketError.Success)
                 {
                     ThreadPool.QueueUserWorkItem(args =>
                     {
-                        var tup = (Tuple<Action<int, byte[], int, SocketError>, int, byte[], int>)args;
+                        var tup = (Tuple<Action<IntPtr, byte[], int, SocketError>, IntPtr, byte[], int>)args;
                         tup.Item1(tup.Item2, tup.Item3, tup.Item4, SocketError.Success);
                     }, Tuple.Create(callback, acceptedFd, socketAddress, socketAddressLen));
                 }
